demo/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
numpy==2.0.2
pgvector==0.3.6
python-dotenv==1.0.1
SQLAlchemy==2.0.38
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
psycopg2==2.9.10
psycopg==3.2.5
psycopg2-binary==2.9.10
gunicorn==20.1.0 
protobuf==6.30.0
tiktoken==0.9.0
sentencepiece==0.2.0
